Implemented SceneNodeMetaData::clone() Fixed some compile warnings Updated gui sln debug config to use jpeg8c, got rid of DEBUG_PATHS.
git-svn-id: file:///home/nick/svnrepo/repo/indigo/trunk@2073 20aa11e4-6e6f-594c-a032-6291ace9ee65

+++ physics/BVH.cpp
@@ -527,7 +527,7 @@ void BVH::doBuild(const AABBox& aabb, std::vector<std::vector<TRI_INDEX> >& tris
 
 	
 	// Alloc space for this node
-	const unsigned int node_index = nodes.size();
+	const unsigned int node_index = (unsigned int)nodes.size();
 	nodes.resize(nodes.size() + 1);
 
 	//assert(node_index < num_nodes);
